move Command type to global types, fix some paths
